+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - (experimental) introduce a new command `bit fork` to export a component to another scope
 
+## [14.2.6-dev.2] - 2019-08-28
+
+- generate node_modules links upon build for new components
+- fix dist replacements upon export (for angular compiler) to support require statements to an internal path
+
 ## [14.2.6-dev.1] - 2019-08-26
 
 - [#1947](https://github.com/teambit/bit/issues/1947) workaround an angular-compiler issue when the dists have a prefix